+++ web/app/themes/welcome.helsinki/app/Blocks/ContentToggle.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace App\Blocks;
+
+use Genero\Sage\NativeBlock\NativeBlock;
+
+class ContentToggle extends NativeBlock
+{
+    public $name = 'hds/content-toggle';
+
+    public function with()
+    {
+        return [
+            'text' => $this->attributes->text,
+        ];
+    }
+}