+++ web/app/themes/welcome.helsinki/resources/assets/scripts/editor.js
@@ -8,6 +8,7 @@ import {
 import { createHigherOrderComponent} from '@wordpress/compose';
 import { addFilter } from '@wordpress/hooks';
 
+import './editor/blocks/content-toggle';
 import './editor/blocks/icon-and-text';
 import './editor/blocks/highlight';
 import './editor/blocks/link';