+++ web/app/themes/welcome.helsinki/resources/assets/scripts/editor/blocks/content-toggle/save.js
@@ -1,0 +1,11 @@
+import { InnerBlocks } from '@wordpress/block-editor'
+
+const save = () => {
+  return (
+    <>
+      <InnerBlocks.Content />
+    </>
+  )
+}
+
+export {save}