+++ web/app/themes/welcome.helsinki/resources/views/blocks/content-toggle.blade.php
@@ -1,0 +1,12 @@
+<div class="{{ esc_attr($block->classes) }}">
+  <div class="wp-block-hds-content-toggle__toggle">
+    <div class="wp-block-hds-content-toggle__text">
+      {{ $text }}
+    </div>
+    <div class="wp-block-hds-content-toggle__icon hds-icon hds-icon--angle-down">
+    </div>
+  </div>
+  <div class="wp-block-hds-content-toggle__content">
+    {!! $content !!}
+  </div>
+</div>